pygeoapi/requirements.txt
Babel
click
filelock
Flask
importlib_metadata
jinja2
jsonschema
pydantic
pygeofilter
pygeoif
pyproj
python-dateutil
pytz
PyYAML
rasterio
requests
shapely<2.0
SQLAlchemy<2.0.0
tinydb
typing-extensions
unicodecsv
gdal==3.6.2
flask-cors
asyncio
nest-asyncio
aiohttp
async-lru
gunicorn
uvicorn
pydash
pillow
python-socketio
